implement facebook login and logout (with ui)
we'll be creating users in the db at this point too.

since authentications belong to user and rails 5 guarantees
that by default there shouldn't be any orphaned authentications
(i.e. authentication records that have user_id as nil) then we
can be sure that each authentication record will have a corresponding
user.

that also means that if we can't find a particular authentication,
then that means the associated user is also not found, so we need
to create a new one to attach to this newly discovered authentication.